Fix Runner loading and refactor code (#977)
**Pull Request Checklist**
- [x] Fixes #962 
- [x] Tests added
- [ ] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**Description of PR**
Primarily, this PR refactors the runner code. It also fixes the mapping
and loading of kwargs to a RunnerInput object.

Currently, the runner code is hard to follow. This PR refactors the
functionality in the `runner.py` file into a `_runner` module with util
submodules. The PR also makes the complex input mapping logic more
testable.

---------
