+++ docs/examples/workflows/template_level_volume.md
@@ -73,35 +73,17 @@ See https://argo-workflows.readthedocs.io/en/latest/walk-through/volumes/
     spec:
       entrypoint: generate-and-use-volume
       templates:
-      - name: generate-and-use-volume
-        steps:
-        - - name: generate-volume
-            template: generate-volume
-            arguments:
-              parameters:
-                - name: pvc-size
-                  # In a real-world example, this could be generated by a previous workflow step.
-                  value: '1Gi'
-        - - name: generate
-            template: whalesay
-            arguments:
-              parameters:
-                - name: pvc-name
-                  value: '{{steps.generate-volume.outputs.parameters.pvc-name}}'
-        - - name: print
-            template: print-message
-            arguments:
-              parameters:
-                - name: pvc-name
-                  value: '{{steps.generate-volume.outputs.parameters.pvc-name}}'
-
-      - name: generate-volume
-        inputs:
+      - inputs:
           parameters:
-            - name: pvc-size
+          - name: pvc-size
+        name: generate-volume
+        outputs:
+          parameters:
+          - name: pvc-name
+            valueFrom:
+              jsonPath: '{.metadata.name}'
         resource:
           action: create
-          setOwnerReference: true
           manifest: |
             apiVersion: v1
             kind: PersistentVolumeClaim
@@ -112,42 +94,62 @@ See https://argo-workflows.readthedocs.io/en/latest/walk-through/volumes/
               resources:
                 requests:
                   storage: '{{inputs.parameters.pvc-size}}'
-        outputs:
-          parameters:
-            - name: pvc-name
-              valueFrom:
-                jsonPath: '{.metadata.name}'
-
-      - name: whalesay
-        inputs:
-          parameters:
-            - name: pvc-name
-        volumes:
-          - name: workdir
-            persistentVolumeClaim:
-              claimName: '{{inputs.parameters.pvc-name}}'
-        container:
+          setOwnerReference: true
+      - container:
+          args:
+          - echo generating message in volume; cowsay hello world | tee /mnt/vol/hello_world.txt
+          command:
+          - sh
+          - -c
           image: docker/whalesay:latest
-          command: [sh, -c]
-          args: ["echo generating message in volume; cowsay hello world | tee /mnt/vol/hello_world.txt"]
           volumeMounts:
-          - name: workdir
-            mountPath: /mnt/vol
-
-      - name: print-message
+          - mountPath: /mnt/vol
+            name: workdir
         inputs:
-            parameters:
-              - name: pvc-name
+          parameters:
+          - name: pvc-name
+        name: whalesay
         volumes:
-          - name: workdir
-            persistentVolumeClaim:
-              claimName: '{{inputs.parameters.pvc-name}}'
-        container:
+        - name: workdir
+          persistentVolumeClaim:
+            claimName: '{{inputs.parameters.pvc-name}}'
+      - container:
+          args:
+          - echo getting message from volume; find /mnt/vol; cat /mnt/vol/hello_world.txt
+          command:
+          - sh
+          - -c
           image: alpine:latest
-          command: [sh, -c]
-          args: ["echo getting message from volume; find /mnt/vol; cat /mnt/vol/hello_world.txt"]
           volumeMounts:
-          - name: workdir
-            mountPath: /mnt/vol
+          - mountPath: /mnt/vol
+            name: workdir
+        inputs:
+          parameters:
+          - name: pvc-name
+        name: print-message
+        volumes:
+        - name: workdir
+          persistentVolumeClaim:
+            claimName: '{{inputs.parameters.pvc-name}}'
+      - name: generate-and-use-volume
+        steps:
+        - - arguments:
+              parameters:
+              - name: pvc-size
+                value: 1Gi
+            name: generate-volume
+            template: generate-volume
+        - - arguments:
+              parameters:
+              - name: pvc-name
+                value: '{{steps.generate-volume.outputs.parameters.pvc-name}}'
+            name: generate
+            template: whalesay
+        - - arguments:
+              parameters:
+              - name: pvc-name
+                value: '{{steps.generate-volume.outputs.parameters.pvc-name}}'
+            name: print
+            template: print-message
     ```
